post_boot: Set reap_mem_on_sigkill only for ulmk
This option should be disabled when KLMK is enabled. The main benefit
is that it will reduce potentially misleading logging information
from the oom_reaper.

+++ rootdir/etc/init.qcom.post_boot.sh
@@ -3537,13 +3537,16 @@ case "$target" in
 	echo 120 > /sys/module/cpu_boost/parameters/input_boost_ms
 
         echo 120 > /proc/sys/vm/watermark_scale_factor
-        echo 1 > /proc/sys/vm/reap_mem_on_sigkill
 
         echo 0-3 > /dev/cpuset/background/cpus
         echo 0-3 > /dev/cpuset/system-background/cpus
 
         # Enable oom_reaper
-        echo 1 > /sys/module/lowmemorykiller/parameters/oom_reaper
+	if [ -f /sys/module/lowmemorykiller/parameters/oom_reaper ]; then
+		echo 1 > /sys/module/lowmemorykiller/parameters/oom_reaper
+	else
+		echo 1 > /proc/sys/vm/reap_mem_on_sigkill
+	fi
 
 	# Enable bus-dcvs
 	for device in /sys/devices/platform/soc
